Try to reduce file length. Add size pref.
